chore: small restructure for better clarity
