+++ tests/service/controllers/test_remote_project.py
@@ -22,7 +22,7 @@ from marshmallow import ValidationError
 import renku
 from renku.core.commands.migrate import migrations_check, migrations_versions
 from renku.core.errors import ConfigurationError
-from renku.service.controllers.utils import RemoteProject
+from renku.service.controllers.utils.remote_project import RemoteProject
 
 
 def test_project_metadata_remote():
